*LINUX PORT* - removed the Tests directory - removed all autotools-related scripts - added simple configuration files to compile autogrid4 and testdiff via CMake under Linux, see INSTALL - snscanf->sscanf (because snscanf is windows-only) - fixed a few g++ compile errors - the auxiliary test script ported to bash

+++ autogrid/trunk/autogrid-4.0.0/utils.cpp
@@ -139,7 +139,7 @@ void endTimer()
     clock_t time = times(&_t) - timers[indexOfNesting];
     for (unsigned int i = 0; i < indexOfNesting; i++)
         fprintf(stderr, "  ");
-    fprintf(stderr, "} took %i ms.\n", (time*1000/getClocksPerSec()));
+    fprintf(stderr, "} took %i ms.\n", int(time*1000/getClocksPerSec()));
 }
 
 // Dummy graphics API entry points.  This app does not do graphics, but it still must provide these callbacks.
